New Editor
Started modifying the new editor to incorporate desired elements presents in Google Docs and Word such as the toolbar always on top, creating a new paragraph creates a new text box and to delete a text, simply remove all the text in it.

There is still a lot of work to be done especially when it comes to redesigning the database so the project isn't fully functional as of now.