Converting the nodejs backend in typescript
I really wanted to convert the backend to typescript to match the frontend with Angular for a while. I tried using the new deno.js from the same designer as node.js but, because it is a new system, there's a lot of bugs and there's few community features.

By using the regular TypeScript system, you can use the technology to compile typescript into regular javascript in node.js. So, I started converting the backend into typescript. Using typescript solves a lot of problems I was having in regular javascript. By having a strong typing, every variables is explicitly defined and many file's data remains active.

I tried to use Socketio and it works for the most part but I got to rethink the new architecture with the new possibilities enabled by TS.

There is still a lot of work to do but it seems promising

Research : 
 - Paper.js
 - Sails
 - Webpack
 - GraphQL
 - Feathers.js