Implemented Concept Entity
A concept represents an idea and, now, a corpus is a collection of documents produced by a single user on a concept. Let's say the concept is "theory of programming languages". A user can add multiple documents explaining the different concepts of that subjects such as "introduction", "functional language", "procedural" and others.